fix bugs in playingScreen and board
