+++ public/index.html
@@ -30,6 +30,7 @@
   height: 100px;
   min-width: 100px;
   min-height: 100px;
+  z-index: 200
 }
 .circular {
   animation: rotate 2s linear infinite;
